requirements.txt
typer
rich
watchdog
pyudev; sys_platform == 'linux'
wmi; sys_platform == 'win32'
pywin32; sys_platform == 'win32'
platformdirs
pyyaml
psutil
